Implement app management support (suspend, resume, close, etc) (#6322)

+++ src/core/hle/applets/applet.cpp
@@ -70,7 +70,7 @@ ResultCode Applet::Create(Service::APT::AppletId id, Service::APT::AppletId pare
     }
 
     Service::APT::AppletAttributes attributes;
-    attributes.applet_pos.Assign(static_cast<u32>(Service::APT::AppletPos::AutoLibrary));
+    attributes.applet_pos.Assign(Service::APT::AppletPos::AutoLibrary);
     attributes.is_home_menu.Assign(false);
     const auto lock_handle_data = manager->GetLockHandle(attributes);
